+++ deployment/v2/build-patterns.sh
@@ -8,17 +8,6 @@ echo "==========================================================================
 echo "aligning versions and updating package.json for CDK v2..."
 /bin/bash $deployment_dir/align-version.sh
 
-echo "============================================================================================="
-echo "updating Import statements for CDK v2..."
-/bin/bash $deployment_dir/rewrite-imports.sh
-
-echo "============================================================================================="
-echo "building cdk-integ-tools..."
-cd $source_dir/tools/cdk-integ-tools
-npm install
-npm run build
-npm link
-
 bail="--bail"
 runtarget="build+lint+test"
 cd $source_dir/
@@ -29,6 +18,18 @@ export NODE_OPTIONS="--max-old-space-size=4096 ${NODE_OPTIONS:-}"
 echo "============================================================================================="
 echo "installing..."
 yarn install --frozen-lockfile
+
+echo "============================================================================================="
+echo "updating Import statements for CDK v2..."
+/bin/bash $deployment_dir/rewrite-imports.sh
+
+echo "============================================================================================="
+echo "building cdk-integ-tools..."
+cd $source_dir/tools/cdk-integ-tools
+npm install
+npm run build
+npm link
+cd $source_dir/
 
 echo "============================================================================================="
 echo "building..."